backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
msal==1.26.0
openai==1.12.0
pytest==8.0.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
python-jose[cryptography]==3.3.0
httpx==0.26.0 